pyproject.toml
[project]
name = "ripoti-kwa-siri"
version = "0.0.1"
description = "Prototype anonymous reporting voice agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "livekit-agents[google]~=1.4",
    "livekit-plugins-silero>=0.1.0",
    "livekit-plugins-turn-detector>=0.1.0",
    "livekit-plugins-noise-cancellation>=0.1.0",
    "google-genai>=1.29.0",
    "openai>=1.99.0",
    "pyyaml>=6.0.1",
    "pydantic-settings>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.uv]
package = false
